support vert.x 3.5.0
