+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.github.st-h</groupId>
     <artifactId>vertx-mongo-streams</artifactId>
-    <version>1.2.1</version>
+    <version>1.3.0</version>
 
     <name>Vert.x - MongoDB streams for the async MongoDB driver</name>
     <description>A stream helper to Pump data from a vert.x ReadStream (e.g. HttpServerFileUpload) to MongoDB AsyncInputStream.</description>
@@ -37,7 +37,7 @@
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
         <mongodb.driver.version>3.3.0</mongodb.driver.version>
-        <stack.version>3.2.0</stack.version>
+        <stack.version>3.5.0</stack.version>
     </properties>
 
     <dependencyManagement>
