+++ src/test/groovy/com/github/sth/vertx/mongo/streams/util/IntegrationTestVerticle.groovy
@@ -1,31 +1,31 @@
-package com.github.sth.groovy.vertx.mongo.streams.util
+package com.github.sth.vertx.mongo.streams.util
 
-import com.github.sth.groovy.vertx.mongo.streams.GridFSInputStream
-import com.github.sth.groovy.vertx.mongo.streams.GridFSOutputStream
+import com.github.sth.vertx.mongo.streams.GridFSInputStream
+import com.github.sth.vertx.mongo.streams.GridFSOutputStream
 import com.mongodb.async.SingleResultCallback
 import com.mongodb.async.client.MongoClient
 import com.mongodb.async.client.MongoClients
 import com.mongodb.async.client.MongoDatabase
 import com.mongodb.async.client.gridfs.GridFSBucket
 import com.mongodb.async.client.gridfs.GridFSBuckets
+import io.vertx.core.AbstractVerticle
 import io.vertx.core.AsyncResult
 import io.vertx.core.Future
 import io.vertx.core.Handler
 import io.vertx.core.http.HttpMethod
-import io.vertx.groovy.core.buffer.Buffer
-import io.vertx.groovy.core.http.HttpServer
-import io.vertx.groovy.core.http.HttpServerFileUpload
-import io.vertx.groovy.core.http.HttpServerRequest
-import io.vertx.groovy.core.http.HttpServerResponse
-import io.vertx.groovy.core.streams.Pump
-import io.vertx.groovy.core.streams.WriteStream
-import io.vertx.lang.groovy.GroovyVerticle
+import io.vertx.core.buffer.Buffer
+import io.vertx.core.http.HttpServer
+import io.vertx.core.http.HttpServerFileUpload
+import io.vertx.core.http.HttpServerRequest
+import io.vertx.core.http.HttpServerResponse
+import io.vertx.core.streams.Pump
+import io.vertx.core.streams.WriteStream
 import org.bson.types.ObjectId
 
 /**
  * A simple verticle, which accepts multipart file upload and serves the uploaded files.
  */
-public class IntegrationTestVerticle extends GroovyVerticle {
+public class IntegrationTestVerticle extends AbstractVerticle {
 
     private HttpServer httpServer
     private MongoClient mongoClient
